pyproject.toml
[project]
name = "agentlangchain"
version = "0.1.0"
description = "Un agent lanchain observable via langsmith et langfuse"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.10",
    "langchain-core>=1.2.14",
    "langchain-groq>=1.1.2",
    "langfuse>=3.14.4",
    "langgraph>=1.0.9",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.21",
]
